+++ apps/api/src/app.module.ts
@@ -2,9 +2,11 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AuditModule } from './modules/audit/audit.module';
 import { AuthModule } from './modules/auth/auth.module';
+import { ExperimentsModule } from './modules/experiments/experiments.module';
 import { HealthModule } from './modules/health/health.module';
 import { PrismaModule } from './modules/prisma/prisma.module';
 import { ProjectsModule } from './modules/projects/projects.module';
+import { RunsModule } from './modules/runs/runs.module';
 import { WorkspacesModule } from './modules/workspaces/workspaces.module';
 
 @Module({
@@ -16,6 +18,8 @@ import { WorkspacesModule } from './modules/workspaces/workspaces.module';
     AuditModule,
     WorkspacesModule,
     ProjectsModule,
+    ExperimentsModule,
+    RunsModule,
   ],
 })
 export class AppModule {}